LayoutTests/imported/w3c/web-platform-tests/html/webappapis/dynamic-markup-insertion/opening-the-input-stream/remove-initial-about-blankness.window-expected.txt


FAIL document.open() removes the initial about:blank-ness of the document assert_equals: expected (undefined) undefined but got (string) "Hello world!"
PASS Double-check: without document.open(), Window reuse indeed happens
